+++ R/gridify.R
@@ -1,6 +1,7 @@
 #' Add Grids to Multiple png
 #' 
-#' Read in a directory of png files.  Add grid lines.  Output to directory.
+#' A wrapper function for \code{grid_calc} and \code{plot_grid} used to read in 
+#' a directory of png files.  Add grid lines.  Output to directory.
 #' 
 #' @param path Path to the in directory with the png files.
 #' @param out Path to the out directory.
@@ -21,7 +22,6 @@
 #' @export
 #' @importFrom parallel parLapply makeCluster detectCores stopCluster clusterEvalQ clusterExport
 #' @importFrom tools file_ext
-#' @importFrom reports delete folder
 #' @seealso \code{\link[embodied]{plot_grid}}
 #' @examples
 #' \dontrun{
@@ -32,7 +32,7 @@ gridify <- function(path = ".", out = file.path(path, "out"),
     columns = 20, rows = columns, parallel = FALSE, cores = detectCores()/2,
     width = 480, height = 480, text.color = "gray60", text.size = 3, 
 	grid.color = text.color, ...){
-
+## detect mp4 file extension and use mp4_to_png to generate pngs w/ grids
     if (file.exists(file.path(path, "out"))) {
         message(paste0("\"", file.path(path, "out"), 
             "\" already exists:\nDo you want to overwrite?\n"))
@@ -49,7 +49,7 @@ gridify <- function(path = ".", out = file.path(path, "out"),
     fls <- dir(path)
     fls <-file.path(path, fls)[file_ext(fls) == "png"]
     dat <- grid_calc(columns, rows)
-
+## Add Lapply handling
         invisible(lapply(fls, function(x){
             png(file.path(out, basename(x)), width=width, height=height, ...)
             print(plot_grid(x, grid.data = dat, text.color = text.color,